some reorganizing and reimplementing of the LINK feature and memory abstraction

+++ retroscope-library/src/main/java/retroscope/rql/TypedValue.java
@@ -10,7 +10,7 @@ public interface TypedValue {
 
     double getDoubleVal();
 
-    int getIntVal();
+    long getIntVal();
 
     String getStringVal();
